requirements.txt
streamlit==1.26.0
tensorflow==2.14.0
Pillow==9.3.0
numpy==1.25.0
langchain-community
langchain-core

replicate
